ml_service/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
prophet
joblib
pymongo
python-dotenv

# Testing dependencies
pytest
pytest-cov